% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_stats.R
\name{boot_stats}
\alias{boot_stats}
\title{Basic bootstrap function statistics}
\usage{
boot_stats(x, d)
}
\arguments{
\item{x}{vector of numeric values}

\item{d}{vector of integrs representing indices}
}
\value{
outputs summary statistics for the values in x sleected by d
}
\description{
Basic bootstrap function statistics
}
\examples{
bootstats(mtcars$cyl, c(1,2,3,4,5,6))
}
