% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hist.R
\name{data_hist}
\alias{data_hist}
\title{Numeric Histograms}
\usage{
data_hist(x, rug = TRUE, dim = NULL)
}
\arguments{
\item{x}{this is a dataframe}

\item{rug}{logical value}

\item{dim}{length 2 vector for the dimensions of the output}
}
\value{
outputs histograms for each numeric variable in a data frame, optionally includes a rug plot,
}
\description{
Numeric Histograms
}
\examples{
data_hist(mtcars, rug = FALSE, dim = c(2,2))
}
