% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_na.R
\name{perc_NA}
\alias{perc_NA}
\title{Percent NA}
\usage{
perc_NA(x, count = TRUE)
}
\arguments{
\item{x}{this is a dataframe}

\item{count}{this is a logical}
}
\value{
Returns a value between 0 and 1 representing the percentage of each column that is NA, and an optional output with the total number of rows that have
at least one NA
}
\description{
Percent NA
}
\examples{
perc_NA(mtcars, count = FALSE)
}
