% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm_stats.R
\name{tsm_stats}
\alias{tsm_stats}
\title{Time Series Model Statistics}
\usage{
tsm_stats(model_list)
}
\arguments{
\item{model_list}{list object containing regression objects}
}
\value{
outputs a table of statistics for each model in the list
}
\description{
Time Series Model Statistics
}
\examples{
x = list(model1, model2, model3)
tsm_stats(x)
}
